requirements.txt
echo "flask
torch
transformers
azure-storage-blob" > requirements.txt
